requirements.txt
asgiref==3.5.2
black==22.8.0
certifi==2022.6.15
click==8.1.3
colorama==0.4.5
Django==4.1
django-ckeditor==6.5.0
django-environ==0.9.0
django-js-asset==2.0.0
django-smartfields==1.1.3
django-versatileimagefield==2.2
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
python-dotenv==0.21.0
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2022.2.1
sentry-sdk==1.9.7
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.2
urllib3==1.26.12
